Work on combining generator classes
