--- Exanite.Core.Generators/Constants.cs
@@ -1,8 +1,0 @@
-using System.Collections.Immutable;
-
-namespace Exanite.Core.Generators;
-
-public static class GeneratorConstants
-{
-    public static readonly ImmutableArray<string> VectorComponents = ["X", "Y", "Z", "W"];
-}
