+++ Exanite.Core.Generators/GeneratorConstants.cs
@@ -1,0 +1,31 @@
+using System.Collections.Immutable;
+using Exanite.Core.Generators.Models;
+
+namespace Exanite.Core.Generators;
+
+public static class GeneratorConstants
+{
+    public static readonly ImmutableArray<string> VectorComponents = ["X", "Y", "Z", "W"];
+
+    public static readonly ImmutableArray<ScalarTypeInfo> ScalarTypes =
+    [
+        new()
+        {
+            Type = ScalarType.Float,
+            TypeName = "float",
+            Suffix = "",
+        },
+        new()
+        {
+            Type = ScalarType.Float,
+            TypeName = "fixed",
+            Suffix = "Fixed",
+        },
+        new()
+        {
+            Type = ScalarType.Float,
+            TypeName = "int",
+            Suffix = "Int",
+        },
+    ];
+}
