+++ Exanite.Core.Generators/Generators/RectGenerator.cs
@@ -1,25 +1,122 @@
 using System.Linq;
 using Exanite.CodeGen;
+using Exanite.Core.Io;
 
 namespace Exanite.Core.Generators.Generators;
 
-public abstract class RectGenerator
+public class RectGenerator
 {
-    protected void AppendConstants(IndentedStringBuilder builder, string rectType, string vectorType)
+    public void Run()
+    {
+        for (var componentCount = 2; componentCount <= 3; componentCount++)
+        {
+            var components = GeneratorConstants.VectorComponents.Take(componentCount).ToArray();
+
+            var rectType = $"Rect{componentCount}Int";
+            var vectorType = $"Vector{componentCount}Int";
+
+            var builder = new IndentedStringBuilder();
+            builder.AppendGeneratedCodeHeader();
+
+            builder.AppendLine("using System.Numerics;");
+            builder.AppendLine();
+            builder.AppendLine("namespace Exanite.Core.Numerics;");
+
+            builder.AppendSeparation();
+            using (builder.EnterScope($"public partial record struct {rectType}"))
+            {
+                AppendConstants(builder, rectType, vectorType);
+                AppendFields(builder, vectorType);
+
+                AppendRectCastOperation(builder, "implicit", rectType, $"Rect{componentCount}", $"Vector{componentCount}");
+                AppendRectCastOperation(builder, "explicit", $"Rect{componentCount}", rectType, vectorType);
+
+                AppendCreateOperations(builder, rectType, vectorType);
+                AppendScaleOperation(builder, rectType, vectorType);
+                AppendContainsOperation(builder, vectorType, components);
+            }
+
+            var outputPath = AbsolutePath.WorkingDirectory / "Exanite.Core" / "Numerics" / $"{rectType}.g.cs";
+            outputPath.WriteAllText(builder.ToString());
+        }
+
+        for (var componentCount = 2; componentCount <= 3; componentCount++)
+        {
+            var components = GeneratorConstants.VectorComponents.Take(componentCount).ToArray();
+
+            var rectType = $"Rect{componentCount}";
+            var vectorType = $"Vector{componentCount}";
+
+            var builder = new IndentedStringBuilder();
+            builder.AppendGeneratedCodeHeader();
+
+            builder.AppendLine("using System.Numerics;");
+            builder.AppendLine();
+            builder.AppendLine("namespace Exanite.Core.Numerics;");
+
+            builder.AppendSeparation();
+            using (builder.EnterScope($"public partial record struct {rectType}"))
+            {
+                AppendConstants(builder, rectType, vectorType);
+                AppendFields(builder, vectorType);
+
+                AppendCreateOperations(builder, rectType, vectorType);
+                AppendScaleOperation(builder, rectType, vectorType);
+                AppendContainsOperation(builder, vectorType, components);
+            }
+
+            var outputPath = AbsolutePath.WorkingDirectory / "Exanite.Core" / "Numerics" / $"{rectType}.g.cs";
+            outputPath.WriteAllText(builder.ToString());
+        }
+
+        for (var componentCount = 2; componentCount <= 3; componentCount++)
+        {
+            var components = GeneratorConstants.VectorComponents.Take(componentCount).ToArray();
+
+            var rectType = $"Rect{componentCount}Fixed";
+            var vectorType = $"Vector{componentCount}Fixed";
+
+            var builder = new IndentedStringBuilder();
+            builder.AppendGeneratedCodeHeader();
+
+            builder.AppendLine("using System.Numerics;");
+            builder.AppendLine();
+            builder.AppendLine("namespace Exanite.Core.Numerics;");
+
+            builder.AppendSeparation();
+            using (builder.EnterScope($"public partial record struct {rectType}"))
+            {
+                AppendConstants(builder, rectType, vectorType);
+                AppendFields(builder, vectorType);
+
+                AppendRectCastOperation(builder, "explicit", rectType, $"Rect{componentCount}", $"Vector{componentCount}");
+                AppendRectCastOperation(builder, "explicit", $"Rect{componentCount}", rectType, vectorType);
+
+                AppendCreateOperations(builder, rectType, vectorType);
+                AppendScaleOperation(builder, rectType, vectorType);
+                AppendContainsOperation(builder, vectorType, components);
+            }
+
+            var outputPath = AbsolutePath.WorkingDirectory / "Exanite.Core" / "Numerics" / $"{rectType}.g.cs";
+            outputPath.WriteAllText(builder.ToString());
+        }
+    }
+
+    private static void AppendConstants(IndentedStringBuilder builder, string rectType, string vectorType)
     {
         builder.AppendSeparation();
         builder.AppendLine($"public static readonly {rectType} Zero = default;");
         builder.AppendLine($"public static readonly {rectType} One = FromSize({vectorType}.One);");
     }
 
-    protected void AppendFields(IndentedStringBuilder builder, string vectorType)
+    private static void AppendFields(IndentedStringBuilder builder, string vectorType)
     {
         builder.AppendSeparation();
         builder.AppendLine($"public {vectorType} Offset;");
         builder.AppendLine($"public {vectorType} Size;");
     }
 
-    protected void AppendRectCastOperation(IndentedStringBuilder builder, string castType, string srcRectType, string dstRectType, string dstVectorType)
+    private static void AppendRectCastOperation(IndentedStringBuilder builder, string castType, string srcRectType, string dstRectType, string dstVectorType)
     {
         builder.AppendSeparation();
         using (builder.EnterScope($"public static {castType} operator {dstRectType}({srcRectType} value)"))
@@ -28,7 +125,7 @@ public abstract class RectGenerator
         }
     }
 
-    protected void AppendCreateOperations(IndentedStringBuilder builder, string rectType, string vectorType)
+    private static void AppendCreateOperations(IndentedStringBuilder builder, string rectType, string vectorType)
     {
         builder.AppendSeparation();
         using (builder.EnterScope($"public static {rectType} FromSize({vectorType} size)"))
@@ -54,7 +151,7 @@ public abstract class RectGenerator
         }
     }
 
-    protected void AppendScaleOperation(IndentedStringBuilder builder, string rectType, string vectorType)
+    private static void AppendScaleOperation(IndentedStringBuilder builder, string rectType, string vectorType)
     {
         builder.AppendSeparation();
         using (builder.EnterScope($"public readonly {rectType} Scale({vectorType} scale)"))
@@ -63,7 +160,7 @@ public abstract class RectGenerator
         }
     }
 
-    protected void AppendContainsOperation(IndentedStringBuilder builder, string vectorType, string[] components)
+    private static void AppendContainsOperation(IndentedStringBuilder builder, string vectorType, string[] components)
     {
         builder.AppendSeparation();
         using (builder.EnterScope($"public readonly bool Contains({vectorType} position)"))
