+++ Exanite.Core.Generators/Models/ScalarType.cs
@@ -1,0 +1,8 @@
+namespace Exanite.Core.Generators.Models;
+
+public enum ScalarType
+{
+    Float,
+    Fixed,
+    Int,
+}
