+++ Exanite.Core.Generators/Models/ScalarTypeInfo.cs
@@ -1,0 +1,8 @@
+namespace Exanite.Core.Generators.Models;
+
+public readonly record struct ScalarTypeInfo
+{
+    public required ScalarType Type { get; init; }
+    public required string TypeName { get; init; }
+    public required string Suffix { get; init; }
+}
